cmd/aspell-dictionaries: rework language finding
The existing logic for identifying `language` and [file] `path`
information in the `aspell-dictionaries` command relies on a specific
HTML structure, so it could fail if there are slight differences
(e.g., capitalization changes in HTML tags/attributes, a different
table cell order, etc.).

This reworks the logic to accomplish the same goal in a way that's
less likely to break. It can still fail if:

* The links are no longer in table cells.
* The path changes in a way that the regex won't match (or handle
  correctly).
* The file type changes to something other than a tarball.

That said, this approach is case-insensitive and doesn't rely on
table cell order, so it won't fail for those reasons.