cmd/aspell-dictionaries: raise on error conditions
The `aspell-dictionaries` command doesn't check for errors, so
failures may not be clear. This adds `raise` calls to make it clear
when and why the command isn't working as expected.

Namely, this adds `--fail` to the `curl_output` arguments, so the
command will have a failing exit status if it receives an HTTP error
(e.g., a 404 Not Found response). This allows us to use
`result.status.success?` to ensure that we received a good response.

This also raises if we fail to identify languages from the HTML
content. Without this, the command completes without producing any
output and it's not clear what went wrong.